SpaceSavingRatio now float (#377)
[CSBS] `SpaceSavingRatio` type to float

What this PR does / why we need it
This PR fixing opentelekomcloud/terraform-provider-opentelekomcloud#1832
Tests
=== RUN   TestBackupList
--- PASS: TestBackupList (1.48s)
=== RUN   TestBackupLifeCycle
    common.go:188: Attempting to create ECSv1
    common.go:201: Created ECSv1 instance: d04f82ca-8f5b-4055-9daf-523925fe47ef
    backup_test.go:38: Check if resource is protectable

    backup_test.go:50: Resource is protectable
    backup_test.go:58: Attempting to create CSBS backup

    backup_test.go:79: Created CSBS backup: 25c83277-902f-4930-b62e-e49bdf4119ef
    backup_test.go:62: Attempting to delete CSBS backup: 25c83277-902f-4930-b62e-e49bdf4119ef
    backup_test.go:68: Deleted CSBS backup: 25c83277-902f-4930-b62e-e49bdf4119ef
    common.go:207: Attempting to delete ECSv1: d04f82ca-8f5b-4055-9daf-523925fe47ef
    common.go:224: ECSv1 instance deleted: d04f82ca-8f5b-4055-9daf-523925fe47ef
--- PASS: TestBackupLifeCycle (362.83s)
PASS

Process finished with the exit code 0

+++ openstack/csbs/v1/backup/results.go
@@ -67,7 +67,7 @@ type ExtendInfo struct {
 	ResourceName         string         `json:"resource_name"`
 	ResourceType         string         `json:"resource_type"`
 	Size                 int            `json:"size"`
-	SpaceSavingRatio     int            `json:"space_saving_ratio"`
+	SpaceSavingRatio     float32        `json:"space_saving_ratio"`
 	VolumeBackups        []VolumeBackup `json:"volume_backups"`
 	FinishedAt           string         `json:"finished_at"`
 	TaskId               string         `json:"taskid"`
